make field_logic.get_next_q_identifier register independent of reset signal
If an asynchronous reset is employed, the value of field_logic.get_next_q_identifier gets updated when the reset signal changes. When the reset signal is asynchronous to the clock, this causes an unwanted update in the register. This change makes the register independent of the reset signal (removes the reset signal from its sensitivity list).